Expected Loss Commit
This is the second part of this project where I attempt to use Stochastic Gradient Descent to predict the Expected Loss from a loan default. 
Features such as probability of default, loss given default, recovery rate and exposure at default are all included in the model when training it on the training data.